inbound-dps-handler-code-jar/src/main/java/com/ericsson/nms/mediation/component/dps/common/SyncType.java
/*------------------------------------------------------------------------------
 *******************************************************************************
 * COPYRIGHT Ericsson 2014
 *
 * The copyright to the computer program(s) herein is the property of
 * Ericsson Inc. The programs may be used and/or copied only with written
 * permission from Ericsson Inc. or in accordance with the terms and
 * conditions stipulated in the agreement/contract under which the
 * program(s) have been supplied.
 *******************************************************************************
 *----------------------------------------------------------------------------*/

package com.ericsson.nms.mediation.component.dps.common;

/**
 * Represents the type of the sync to be performed.
 */
public enum SyncType {

    DELTA("delta"), FULL("full");

    private String type;

    SyncType(final String pType) {
        this.setType(pType);
    }

    public String getType() {
        return type;
    }

    public void setType(final String type) {
        this.type = type;
    }

}
